fix: 서버쪽 인증 로직을 따로 분리 및 Mutation으로 변경
user 서비스는 user의 데이터를 처리하는 공간인데 이곳에서 인증을 처리하는 점이 어색하다고 느껴져서 새롭게
auth 서비스에서 인증 로직을 처리하도록 수정함.
아울러 인증 관련해서 요청이 들어오면 디비에 토큰을 저장하는 등의 사이드 이펙트가 발생하므로 Query보다 Mutation이
적절하다고 판단되어 이를 Mutation으로 수정함
