+++ server/src/services/auth/authType.ts
@@ -1,0 +1,9 @@
+export default `
+type Auth {
+  accessToken: String
+}
+
+type Mutation {
+  login(code: String!): Auth!
+}
+`;
